+++ src/app/app.module.ts
@@ -8,7 +8,7 @@ import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { TopHeaderComponent } from './components/top-header/top-header.component';
 import { DashboardComponent } from './components/dashboard/dashboard.component';
-import { LoginComponent } from './components/login/login.component';
+import { LoginComponent } from './components/member/login/login.component';
 import { HomeComponent } from './components/home/home.component';
 import { HelpComponent } from './components/help/help.component';
 import { RegisterComponent } from './components/member/register/register.component';
@@ -16,6 +16,7 @@ import { ForgotPasswordComponent } from './components/member/forgot-password/for
 import { AddHelpComponent } from './components/help/add-help/add-help.component';
 import { AddFundComponent } from './components/help/add-fund/add-fund.component';
 import { BalanceComponent } from './components/member/balance/balance.component';
+
 
 @NgModule({
   declarations: [
